Use `int64_t` for many things, and some other minor cleanups (#2102)
We have been using many different integer types, and it is annoying
because we have to cast. Let's unify it and use `int64_t` everywhere.

+++ csrc/compute_at_map.cpp
@@ -1688,7 +1688,7 @@ void IterDomainGraph::updateComputeWith(TensorView* compute_with_tv) {
   for (auto pos = compute_with_tv->getComputeAtPosition();
        pos < compute_with_tv->getComputeWithPosition();
        ++pos) {
-    auto id = compute_with_tv->axis((int)pos);
+    auto id = compute_with_tv->axis(pos);
 
     // Find the matching consumer ID using the permissive map
     auto it = std::find_if(
